chore: add send bundle trait
